update gaussian profile

+++ plugins/ml/src/main/java/greycat/ml/math/Gaussian1D.java
@@ -17,7 +17,7 @@ package greycat.ml.math;
 
 public class Gaussian1D {
 
-    public static double getCovariance(double sum, double sumSq, int total) {
+    public static double getCovariance(double sum, double sumSq, long total) {
         return (sumSq - (sum * sum) / total) / (total - 1);
     }
